Break /selectTenant into two endpoints
/tenant/choose: pick a tenant when one wasn't specified (or couldn't be defaulted to)
/tenant/unauthorized: pick a different tenant because you don't have access to the current one
