+++ lib/tenant.js
@@ -48,11 +48,11 @@ module.exports = {
   ensureCurrent: function () {
     return function (req, res, next) {
       if (!req.tenant)
-        return res.redirect('/selectTenant');
+        return res.redirect('/tenant/choose');
       
       var tenants = getAuthorizedTenants(req);
       if (!tenants || !tenants.some(tenant => tenant === req.tenant))
-        return res.redirect('/selectTenant');
+        return res.redirect('/tenant/unauthorized');
 
       next();
     };
