+++ views/select_tenant.jade
@@ -3,7 +3,7 @@ extends layout
 block content  
   h1= title
   div
-    p Oh no! I can't figure out where to send you.  Please tell me where you want to go.
+    p= message
     ul
       - each tenant in tenants
         li
